+++ crates/eframe/src/native/epi_integration.rs
@@ -3,6 +3,9 @@ use winit::event_loop::EventLoopWindowTarget;
 #[cfg(target_os = "macos")]
 use winit::platform::macos::WindowBuilderExtMacOS as _;
 
+#[cfg(feature = "wayland")]
+use winit::platform::wayland::WindowBuilderExtWayland as _;
+
 #[cfg(feature = "accesskit")]
 use egui::accesskit;
 use egui::NumExt as _;
@@ -116,6 +119,12 @@ pub fn window_builder<E>(
             .with_title_hidden(true)
             .with_titlebar_transparent(true)
             .with_fullsize_content_view(true);
+    }
+
+    #[cfg(feature = "wayland")]
+    {
+        window_builder =
+            window_builder.with_name(native_options.app_id.as_deref().unwrap_or(title), "");
     }
 
     if let Some(min_size) = *min_window_size {